Setup Testing/requirements.txt
alabaster==0.7.13
Babel==2.13.1
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
docutils==0.20.1
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
numpydoc==1.6.0
packaging==23.2
Pygments==2.16.1
pywin32==306
pywinpty==2.0.12
requests==2.31.0
setuptools==68.2.2
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
tabulate==0.9.0
TDTPy==0.10.2
urllib3==2.0.7